Android Teddy v0.6
-yet to finish
1. portfolio sell
2. news
3. insights
4. peers
5. click on portfolio item to search result

+++ androidteddy/app/src/main/java/com/example/androidteddy/FavoritesSection.java
@@ -47,9 +47,9 @@ public class FavoritesSection extends Section {
         FavoritesItemViewHolder itemView = (FavoritesItemViewHolder) holder;
         itemView.setSymbol((String) itemList.get(position).get("symbol"));
         itemView.setCompanyName((String) itemList.get(position).get("companyName"));
-        itemView.setC((Double) itemList.get(position).get("c"));
-        itemView.setD((Double) itemList.get(position).get("d"));
-        itemView.setDP((Double) itemList.get(position).get("dp"));
+        itemView.setC((Float) itemList.get(position).get("c"));
+        itemView.setD((Float) itemList.get(position).get("d"));
+        itemView.setDP((Float) itemList.get(position).get("dp"));
     }
